passport setup
looks like passport should be just about configured. why was password
select false in the user model? do we send that to the user or
something?

I'm not going to worry about that for now -- it'll come out in a broader
refactoring of how server calls are made. I mean, realistically we need
to split this into an api / client server anyway.