passport redirect works
It's a little bit jenky -- rather than the server redirecting the user,
the server sends a json object to the user and the client redirects. but
it works, and that might be a totally normal way to do it (probably is,
tbh). So the dashboard doesn't work right now, but i think that this is
probably a good opportunity to split the UI and API servers